orchestration-with-cloud-composer-and-airflow/requirements.txt
apache-airflow==1.10.15
black==21.4b0
SQLAlchemy==1.3.23 
markupsafe==2.0.1
wtforms==2.2
google-api-python-client==2.79.0
apache-airflow-backport-providers-mysql==2021.3.3
psycopg2-binary==2.9.5
google-cloud-storage==2.7.0
pandas-gbq==0.14.1
PyMySQL==1.0.2
cloud-sql-python-connector[pymysql]
pylint

